isolate salect_multiple_gfile node interface
